refactored events to all use 32 bit ints (a small boost for some GPUs) and begun work on queueless

+++ src/simulation.cpp
@@ -492,7 +492,7 @@ void finalize_generation()
   }
 }
 
-double initialize_history(Particle& p, int64_t index_source)
+double initialize_history(Particle& p, int index_source)
 {
   // set defaults
   if (settings::run_mode == RunMode::EIGENVALUE) {
